Fix column reordering in Safari and Firefox (#511)
Fixes #427 

Noticed this problem in Firefox as well. I think this pull request
addresses 2 little problems:

1. The `dragover` event gets fired continuously during dragging. On the
callback the DOM is manipulated to update the list. Somehow Chrome
handles this pretty well but on Firefox it just flooded the event queue
I guess and the `drop` event was not fired most of the time. I added a
simple check below to see if we should update the DOM or not.

```javascript
if (draggingElement.id === item.id) return;
```

2. Now Firefox was working, but Safari was still giving problems.
Setting the drag image to an empty div doesn't work in Safari
[apparently](https://stackoverflow.com/a/40923520/943337). I replaced it
with an inline GIF, a 1x1 transparent pixel, [the smallest I could
find](https://stackoverflow.com/a/9967193/943337).

Tested on Firefox, Safari and Chrome. Thanks for your work!